// dashboard/src/client/models/ui-event-envelope.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UiEventPayload } from './ui-event-payload';
// May contain unused imports in some cases
// @ts-ignore
import type { UiEventScope } from './ui-event-scope';
// May contain unused imports in some cases
// @ts-ignore
import type { UiEventType } from './ui-event-type';

export interface UiEventEnvelope {
    'version': UiEventEnvelopeVersionEnum;
    'id': string;
    'type': UiEventType;
    'occurredAt': string;
    'scope': UiEventScope;
    'replayed'?: boolean | null;
    'correlationId'?: string | null;
    'meta'?: { [key: string]: any; } | null;
    'data': UiEventPayload;
}

export const UiEventEnvelopeVersionEnum = {
    V1: 'v1'
} as const;

export type UiEventEnvelopeVersionEnum = typeof UiEventEnvelopeVersionEnum[keyof typeof UiEventEnvelopeVersionEnum];